+++ PACECloudSDK/IDKit/Session/IDKit+API+TokenHandling.swift
@@ -9,8 +9,14 @@ import Foundation
 
 extension IDKit {
     func performApiInducedRefresh(_ completion: @escaping (Bool) -> Void) {
-        performRefresh { accessToken, error in
-            completion(error == nil && accessToken != nil)
+        performRefresh { result in
+            switch result {
+            case .success(let accessToken):
+                completion(accessToken != nil)
+
+            case .failure:
+                completion(false)
+            }
         }
     }
 }